# Improvements to the performance of persisting timeline metadata records.