--- a/src/main/java/org/sagebionetworks/bridge/hibernate/HibernateSchedule2Dao.java
+++ b/src/main/java/org/sagebionetworks/bridge/hibernate/HibernateSchedule2Dao.java
@@ -36,7 +36,7 @@ import org.sagebionetworks.bridge.models.schedules2.timelines.TimelineMetadata;
 
 @Component
 public class HibernateSchedule2Dao implements Schedule2Dao {
-	private static final Logger LOG = LoggerFactory.getLogger(HibernateSchedule2Dao.class);
+    private static final Logger LOG = LoggerFactory.getLogger(HibernateSchedule2Dao.class);
 
     static final String SELECT_COUNT = "SELECT count(*) ";
     static final String GET_ALL_SCHEDULES = "FROM Schedule2 WHERE appId=:appId";
@@ -46,7 +46,10 @@ public class HibernateSchedule2Dao implements Schedule2Dao {
     static final String DELETE_ORPHANED_SESSIONS = "DELETE FROM Sessions where scheduleGuid = :guid AND guid NOT IN (:guids)";
     static final String AND_DELETED = "AND deleted = 0";
     static final String AND_NOT_IN_GUIDS = "AND guid NOT IN (:guids)";
-    static final String INSERT = "INSERT INTO TimelineMetadata (appId, assessmentGuid, assessmentId, assessmentInstanceGuid, assessmentRevision, scheduleGuid, scheduleModifiedOn, schedulePublished, sessionGuid, sessionInstanceEndDay, sessionInstanceGuid, sessionInstanceStartDay, sessionStartEventId, timeWindowGuid, timeWindowPersistent, guid) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)";
+    static final String INSERT = "INSERT INTO TimelineMetadata (appId, assessmentGuid, assessmentId, assessmentInstanceGuid, "
+            + "assessmentRevision, scheduleGuid, scheduleModifiedOn, schedulePublished, sessionGuid, sessionInstanceEndDay, "
+            + "sessionInstanceGuid, sessionInstanceStartDay, sessionStartEventId, timeWindowGuid, timeWindowPersistent, guid) "
+            + "VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)";
     static final String DELETE_TIMELINE_RECORDS = "DELETE FROM TimelineMetadata WHERE scheduleGuid = :scheduleGuid";
     static final String SELECT_ASSESSMENTS_FOR_SESSION_INSTANCE = "SELECT * FROM TimelineMetadata WHERE sessionInstanceGuid = :instanceGuid AND assessmentInstanceGuid IS NOT NULL";
 
@@ -73,7 +76,7 @@ public class HibernateSchedule2Dao implements Schedule2Dao {
     }
 
     @Override
-	public PagedResourceList<Schedule2> getSchedules(String appId, int offsetBy, int pageSize, boolean includeDeleted) {
+    public PagedResourceList<Schedule2> getSchedules(String appId, int offsetBy, int pageSize, boolean includeDeleted) {
         checkNotNull(appId);
 
         QueryBuilder query = new QueryBuilder();
@@ -81,8 +84,8 @@ public class HibernateSchedule2Dao implements Schedule2Dao {
         if (!includeDeleted) {
             query.append(AND_DELETED);
         }
-        List<Schedule2> results = hibernateHelper.queryGet(query.getQuery(), query.getParameters(), 
-        		offsetBy, pageSize, Schedule2.class);
+        List<Schedule2> results = hibernateHelper.queryGet(query.getQuery(), query.getParameters(), offsetBy, pageSize,
+                Schedule2.class);
 
         int total = hibernateHelper.queryCount(SELECT_COUNT + query.getQuery(), query.getParameters());
 
@@ -101,8 +104,8 @@ public class HibernateSchedule2Dao implements Schedule2Dao {
             query.append(AND_DELETED);
         }
 
-        List<Schedule2> results = hibernateHelper.queryGet(query.getQuery(), query.getParameters(), 
-        		offsetBy, pageSize, Schedule2.class);
+        List<Schedule2> results = hibernateHelper.queryGet(query.getQuery(), query.getParameters(), offsetBy, pageSize,
+                Schedule2.class);
 
         int total = hibernateHelper.queryCount(SELECT_COUNT + query.getQuery(), query.getParameters());
 
@@ -141,34 +144,35 @@ public class HibernateSchedule2Dao implements Schedule2Dao {
         // Update the schedule
         hibernateHelper.executeWithExceptionHandling(schedule, (session) -> {
             Set<String> sessionGuids = schedule.getSessions().stream().map(Session::getGuid).collect(toSet());
-            
+
             // Although orphanRemoval = true is set for the session collection, they do not
-            // delete when removed. So we are manually finding the removed sessions and deleting
+            // delete when removed. So we are manually finding the removed sessions and
+            // deleting
             // them before persisting the new set.
             QueryBuilder builder = new QueryBuilder();
             builder.append(DELETE_SESSIONS, GUID, schedule.getGuid());
             if (!sessionGuids.isEmpty()) {
                 builder.append(AND_NOT_IN_GUIDS, GUIDS, sessionGuids);
             }
-        	NativeQuery<?> query = session.createNativeQuery(builder.getQuery());
+            NativeQuery<?> query = session.createNativeQuery(builder.getQuery());
             for (Map.Entry<String, Object> entry : builder.getParameters().entrySet()) {
                 query.setParameter(entry.getKey(), entry.getValue());
             }
             query.executeUpdate();
 
             session.update(schedule);
-            
+
             return schedule;
         });
         deleteAndRecreateTimelineMetadataRecords(schedule, true);
         return schedule;
     }
-    
+
     private void deleteAndRecreateTimelineMetadataRecords(Schedule2 schedule, boolean deleteFirst) {
         Timeline timeline = Scheduler.INSTANCE.calculateTimeline(schedule);
         List<TimelineMetadata> metadata = timeline.getMetadata();
 
-    	hibernateHelper.executeWithExceptionHandling(schedule, (session) -> {
+        hibernateHelper.executeWithExceptionHandling(schedule, (session) -> {
             // batch these operations. Improves network performance
             session.setJdbcBatchSize(batchSize);
 
@@ -178,50 +182,55 @@ public class HibernateSchedule2Dao implements Schedule2Dao {
                 query.setParameter(SCHEDULE_GUID, schedule.getGuid());
                 query.executeUpdate();
                 deleteMetadataStopwatch.stop();
-                
-                LOG.info("Batch deleting timeline metadata records in " + deleteMetadataStopwatch.elapsed(MILLISECONDS) + " ms");
+
+                LOG.info("Batch deleting timeline metadata records in " + deleteMetadataStopwatch.elapsed(MILLISECONDS)
+                        + " ms");
             }
 
-            // Hibernate’s session.save() does an insert and then an update operation on each record, so switching
-            // to JDBC to do an insert only, halves the time it takes to do this operation even without further batch
-            // optimizations. I was not able to determine why Hibernate is doing this (it’s not the most frequently 
+            // Hibernate’s session.save() does an insert and then an update operation on
+            // each record, so switching
+            // to JDBC to do an insert only, halves the time it takes to do this operation
+            // even without further batch
+            // optimizations. I was not able to determine why Hibernate is doing this (it’s
+            // not the most frequently
             // cited culprit, an @Id generator, because we don’t use one).
             Stopwatch createMetadataStopwatch = Stopwatch.createStarted();
             session.doWork(persistRecordsInBatches(metadata));
             createMetadataStopwatch.stop();
-            
+
             LOG.info("Persisting " + metadata.size() + " timeline metadata records in "
                     + createMetadataStopwatch.elapsed(MILLISECONDS) + " ms (batchSize = " + batchSize + ")");
 
             return null;
-        });    	
+        });
     }
 
     /**
-     * For batch operations to work efficiently using the MySQL driver, rewriteBatchedStatements=true must be 
-     * included in the connector string, auto commit must be off, and you must use the batch commit method. 
-     * A batch size of 100 seems about optimal (values below lose performance but I cannot measure any benefit 
-     * to having larger values). 
+     * For batch operations to work efficiently using the MySQL driver,
+     * rewriteBatchedStatements=true must be included in the connector string, auto
+     * commit must be off, and you must use the batch commit method. A batch size of
+     * 100 seems about optimal (values below lose performance but I cannot measure
+     * any benefit to having larger values).
      */
-	protected Work persistRecordsInBatches(List<TimelineMetadata> metadata) {
-		return (connection) -> {
-		    connection.setAutoCommit(false);
-		    PreparedStatement ps = connection.prepareStatement(INSERT);
-		    
-		    for (int i = 0, len = metadata.size(); i < len; i++) {
-		        TimelineMetadata meta = metadata.get(i);
-		        updatePreparedStatement(ps, meta);
-		        if (i > 0 && (i % batchSize) == 0) {
-		        	ps.executeBatch();
-		        }
-		    }
-		    ps.executeBatch();
-	    };
-	}
-	
-	// For testability, removing this to a separate method
-	protected void updatePreparedStatement(PreparedStatement ps, TimelineMetadata meta) throws SQLException {
-		ps.setString(1, meta.getAppId());
+    protected Work persistRecordsInBatches(List<TimelineMetadata> metadata) {
+        return (connection) -> {
+            connection.setAutoCommit(false);
+            PreparedStatement ps = connection.prepareStatement(INSERT);
+
+            for (int i = 0, len = metadata.size(); i < len; i++) {
+                TimelineMetadata meta = metadata.get(i);
+                updatePreparedStatement(ps, meta);
+                if (i > 0 && (i % batchSize) == 0) {
+                    ps.executeBatch();
+                }
+            }
+            ps.executeBatch();
+        };
+    }
+
+    // For testability, removing this to a separate method
+    protected void updatePreparedStatement(PreparedStatement ps, TimelineMetadata meta) throws SQLException {
+        ps.setString(1, meta.getAppId());
         ps.setString(2, meta.getAssessmentGuid());
         ps.setString(3, meta.getAssessmentId());
         ps.setString(4, meta.getAssessmentInstanceGuid());
@@ -229,7 +238,7 @@ public class HibernateSchedule2Dao implements Schedule2Dao {
         if (meta.getAssessmentRevision() == null) {
             ps.setNull(5, Types.NULL);
         } else {
-            ps.setInt(5, meta.getAssessmentRevision());    
+            ps.setInt(5, meta.getAssessmentRevision());
         }
         ps.setString(6, meta.getScheduleGuid());
         ps.setLong(7, meta.getScheduleModifiedOn().getMillis());
@@ -242,9 +251,9 @@ public class HibernateSchedule2Dao implements Schedule2Dao {
         ps.setString(14, meta.getTimeWindowGuid());
         ps.setBoolean(15, meta.isTimeWindowPersistent());
         ps.setString(16, meta.getGuid());
-        ps.addBatch();		
-	}
-    
+        ps.addBatch();
+    }
+
     @Override
     public void deleteSchedule(Schedule2 schedule) {
         checkNotNull(schedule);
@@ -277,14 +286,15 @@ public class HibernateSchedule2Dao implements Schedule2Dao {
         TimelineMetadata tm = hibernateHelper.getById(TimelineMetadata.class, instanceGuid);
         return Optional.ofNullable(tm);
     }
-    
+
     @Override
     public List<TimelineMetadata> getAssessmentsForSessionInstance(String instanceGuid) {
         checkNotNull(instanceGuid);
-        
+
         QueryBuilder builder = new QueryBuilder();
         builder.append(SELECT_ASSESSMENTS_FOR_SESSION_INSTANCE, INSTANCE_GUID, instanceGuid);
-        
-        return hibernateHelper.nativeQueryGet(builder.getQuery(), builder.getParameters(), null, null, TimelineMetadata.class);
+
+        return hibernateHelper.nativeQueryGet(builder.getQuery(), builder.getParameters(), null, null,
+                TimelineMetadata.class);
     }
 }