… src/main/java/org/sagebionetworks/bridge/validators/ValidatorUtils.java
… src/main/resources/BridgeServer2.conf
… src/test/java/org/sagebionetworks/bridge/hibernate/HibernateSchedule2DaoTest.java
… src/test/java/org/sagebionetworks/bridge/validators/Schedule2ValidatorTest.java
… src/test/java/org/sagebionetworks/bridge/validators/ValidatorUtilsTest.java