See BRIDGE-3078:

1. Added an index for TimelineMetadata scheduleGuid, although this had a marginal impact on the call time because the deletion was not the majority of the time being spent on the call (it was about 10 seconds of deleting and over a minute of record insertions).
2. Switched to JDBC to avoid a Hibernate behavior of inserting and then updating each record, which I could not prevent. This halved insert time;
3. Properly implmented batch calls by rewriting insert statements to concatenate them in the SQL, which halved the insert time again (times reduced to approximately 5s and 15s, respectively);
4. Used validation to limit the duration of a schedule to five years (the example that triggered this issue was ten years long—no study that has been proposed that I’m aware of has been ten years long). This brings the total call time down to under 3s, which is acceptable.